Fixed up pricing rules!
