+++ src/main/scala/org/economicsl/auctions/multiunit/pricing/package.scala
@@ -13,11 +13,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.economicsl.auctions.pricing
-
-import org.economicsl.auctions.{LimitAskOrder, LimitBidOrder, Price, Tradable}
+package org.economicsl.auctions.multiunit
 
 
-/** Discriminatory pricing rule computes a different price for each pair of orders. */
-trait DiscriminatoryPricingRule[T <: Tradable] extends (((LimitAskOrder[T], LimitBidOrder[T])) => Price)
-
+/** Documentation for the pricing package should go here! */
+package object pricing
